"Peek" panel for 1.5 seconds when highlighted (#8659)
* panel.js: peek hidden panels on highlight

Panels are highlighted by cs_panel.py and cs_applet.py to indicate
which panel you're working on. With hidden panels, it can be very
difficult to notice as only a 1 pixel strip is visible.

This makes a hidden panel show for PEEK_PANEL_TIME milliseconds
when highlight is turned on. Currently 1500ms.

* panel.js: rework _updatePanelVisibility a bit

To optimize for panel edit mode and peeking cases, especially when
using intelligent autohide.