Menu: Tune search results
Fixes #8070

Takes elements from https://github.com/linuxmint/cinnamon/pull/8251

In the absence of an exact match on app name, any match of the pattern in the description,
keyword list, app ID or app name was enough. This is no longer the case.

- For the App ID to be a match, it must start with the pattern, not just contain it. For instance,
if you search for "mint" it will return "mintlocale", but if you search for "loc", it won't.

- For the App Name to be match, one of its words must start with the pattern. For instance,
"fox" would have previously shown "Firefox" as a hit, it no longer does.

- Same thing for the description, one of its words must start with the pattern.

- It no longer matters if the keywords list contains the pattern, one of the keywords itself
now must start with the pattern. "draw" will return Inkscape, but "torial" won't.