Cinnamon Settings: Fix search for locales with a non-Latin character set
Currently all characters in a search string which can not be represented in the
ASCII character set filtered from the search string. This happens to work for
Latin based locales because the strings first get normalized to NFD, which
represents e.g. 'é' as the base character 'e' followed by the combining
character '´'. Only the combining characters get filtered and the base
characters remain.

However, in locales, which do not use the Latin character set, the base
characters get filtered as well and the search string ends up being empty.

Fix this by not converting to ASCII and instead filtering the combining
characters explicitly.

Fixes https://github.com/linuxmint/cinnamon/issues/8662