Cinnamon-0.1.typelib: do not link to libnm typelib
Nothing in libcinnamon uses libnm, and the typelib does not need to
either. In fact, the only part of cinnamon that uses libnm is the applet
network@cinnamon.org and that directly imports.gi.NM, so it doesn't need
to get it via exposure through the GObject bindings for cinnamon itself.

Removing references to NM from the Cinnamon typelib and rebuilding
cinnamon results in a fully working Cinnamon desktop and a fully working
network applet, so there is no downside to just dropping it.