menu applet: remove some pointless state tracking
initial_load_done was unused, and menuIsOpening was used to stop
search from running early, but running on the 2nd changed event
seems pretty random. Instead just don't run search if the text is
empty, or the hint text(if that can even happen). If we have
partially loaded results ever, one keystroke corrects it but
chances are that someone can't even type a full query while the
menu is opening.

A more correct implementation of this behavior would be to prevent
searching during any _clearAllSelections(), _displayButtons(), or
_refresh*() functions as well as during opening (until initial_cat
load() is done). These alone are synchronous, so should be "atomic"
from the point of view of a search callback running, but usually
multiple steps are needed to complete a full refresh and we use
timeouts and idles here and there. Best bet is to lock searching
during any refresh or display state change for the entire operation.

That said, I haven't seen a need to fix this as menu-open refreshes
are rare, and as previously mentioned typing a full query during
initial opening is unlikely. Refreshes automatically reset the
menu state when they are done, and a keystroke corrects the
premature search condition.