cinnamon-window-tracker.c: use screen signals instead of workspace
This fixes apps not being associated yet for js screen signal
listeners, avoids redundant retrackings when a window changes
workspace, and avoids re-connecting signals when the number of
workspaces changes.

This also makes sure we only emit one signal per change, and don't
emit at all during init. There doesn't really seem to be a point to
emit at that point as there would be no connections and we aren't
using any default handlers.

Since we now track windows for their entire lifetime, we no longer
disassociate the window from the app on window-removed. There is a
window::unmanaged listener in CinnamonApp for that time. We still
need to do this when we retrack for a property change, though.