menu applet: refreshless and more consistent icon settings
refreshless:
Buttons now always add an icon and the visibility is set
based upon the setting, instead of adding it when the
setting is on and rebuilding when it is toggled.

recent clear button:
Text is now bold so it's more easily distinguishable when
application icons are turned off, and it now respects the
show application icons setting.

recent category:
Now properly toggles icon when the setting is turned off
after previously being added to the list with the setting
on.

temporary buttons:
Now respect the show application icons setting and don't add
an icon if it is disabled. Since they only last a short time,
they don't support re-showing icons like normal buttons do.