menu applet: finish adding SMI type strings
_displayButtons() now supports specifying an app categoryId or
SimpleMenuItem type string, or a list of buttons to show.

categoryId and type strings can be partial. The only current
use of this functionality is for the following case:
"recent" will match the "no-recent" and "recent-clear" button
types.

This also removes the redundant listBookmarks and Devices functions
and additionally uses the getAllPlaces() function instead which
includes default bookmarks and additional places features like
"connect to..."