menu applet: rework refreshing
This reworks refreshing so that you queue a specific refresh type
(or multiple) at any point and don't worry about whether the rest
of the categories physically below it need to be refreshed as well.
We now preserve menu order by inserting each button type directly
to the relative index it should be at. For example: a place button
should be inserted after the last application button.

This also queues all refreshes over 1 second from the first, and
runs all queued refreshes at once in the correct order. As a result,
some refreshes that were instant before are delayed now.

Finally, most button actors are now hidden by default. This simplifies
the visibility management with random refreshes. We now only ever
have to hide actors that we have previously shown. Previously, if a
refresh happened while the menu is closed we would have extra actors
in the all apps category (places and recent). This is due to a optimized
initial state set code that doesn't do all the work _displayButtons()
does. One downside of this change is that if the menu refreshes
during use, then buttons may completely disappear until a search
or category select.