menu applet: rework refreshing
This reworks refreshing so that you queue a specific refresh type
(or multiple) at any point and don't worry about whether the rest
of the categories physically below it need to be refreshed as well.
We now preserve menu order by inserting each button type directly
to the relative index it should be at. For example: a place button
should be inserted after the last application button.

This also delays all refreshes by 500ms each time there is a new
one queued before the last one has run. Once 500ms has elapsed, then
all queued refreshes are run at once in the correct order. As a
result, some refreshes that were instant before are delayed now as
long as refresh triggers are ongoing.

Finally, most button actors are now hidden by default if the menu is
not open. This simplifies visibility management with random refreshes.
Previously, if a refresh happened while the menu is closed we would have
extra actors in the all apps category (places and recent). This is due to
an optimized initial state set code that doesn't do all the work
_displayButtons() does. One downside of this change is that if the menu
refreshes during use, then buttons may still appear in the wrong category
as they can now.