cs_themes.py: speed up loading of the icon for the icon theme widget
This speeds up the loading of the icon theme widget (and therefore the entire module) by keeping track of the path to the folder icon and storing it in a file for the next use. Local tests showed a drop in load time (icon themes only) from 1.2s-1.7s to ~.1s with disk cache and 2.4s down to .9s without disk cache.