Window settings: Simply layout settings
The previous settings caused the following issues:

- They were niche/overkill
- They advertised features (pin/shade) which aren't always properly supported by
Cinnamon and/or desktop themes
- They made the setting window really wide

A vast majority of users keep their buttons on the right. There's a minority coming
from Mac or Ubuntu which is used to having them on the left. Other than that, we're
getting really niche.

If there's a lot of demand for another layout we can add it as a combo option.

Niche cases also have the ability of fine tuning the layout via gsettings.