settings.js: allow distributions to override individual settings keys
The override file has the same format as the settings-schema.json file, except that only the overridden settings should be included. The file must be placed in the applet folder, and have the name `settings-override.json`. Just as with settings-schema.json, changes to (including creation or deletion of) this file will automatically trigger an upgrade of the settings.

Some additional improvements to the code were implemented for consistency and better error handling and detection.