menu-applet: don't show user-hidden categories
I inadvertently removed this feature in the refactoring. User-
hidden directories are marked nodisplay and so that needs to be
checked on the gmenu object either at load or display. This commit
does it at load, so consequently any unhidden category will show
all unique apps as newly installed/highlighted upon unhiding after
starting the session or applet with it hidden.

Thanks goes to @collinss for finding both the bug and the fix.