window-list applet: Remove window-actor check during window-monitor- changed.
This was incorrectly returning and not updating for the new monitor,
causing the window list item to end up on a different monitor than
the window (in multimonitor, multi-window-list setups).

This should be looked at again at some point, the window-monitor-
changed signal tends to get fired before the async callback for
window-added.  In single monitor setups, the window never changes
after spawning.  The window briefly appears on the left-most monitor
rather than the primary (or whatever monitor it was launched from)
if the two are not the same.  This seems like some issue in muffin.
This didn't occur prior to 4.0.