window-list: proper fix for incorrect flashing (#8968)
The original implementation checked needsAttention every timeout
interval and this caught the case where a window demands attention
event is triggered by an activate event on an existing window. An
example of this could be a non-focused browser window being focused
as a user clicks a link in another application. In this case, with
the original implementation, the window starts flashing and stops
before the style even changes. It also handles the case where a
user clicks the window during the flash cycle.

This commit restores that behavior.

fixes/reverts: ec9b3fb1c1c03eaf27b893e0f2ffc29166b646c1