grouped-window-list@cinnamon.org: Listen to MetaDisplay's window-created signal for adding new windows.
MetaWorkspace:window-added is sent very early in the construction of a new
window - before the compositor even knows about it, which ends up with cinnamon
frequently adding a window with no actor yet (to clone for thumbnails).

The display signal is sent at the very end of meta_window_new_with_attrs() so
should be more reliable to provide a fully developed window.

If a window is created using the ctr-click method on the app's panel item,
the framing around the new thumb is corrupted at first, if animation is enabled.
Removing the hide/show when adding a new window in this manner avoids this issue,
and has the added benefit of the entire thumbnail group re-centering itself with
its panel item automatically.