cs_themes.py: Change how the gtk3 css override file is handled.
The tinycss parser was being used exclusively to read and then write
back contents of the file.  This would result in anything in the file
that wasn't a "Ruleset" either breaking things, or being discarded
(comments).

This will now read the entire contents of the file, and when re-writing,
preserve all lines except any pre-existing for the scrollbar slider,
and append the slider rule to the end of the file.

Fixes #9043 as now we only concern ourselves with a single ruleset,
of a known type that is supported by tinycss.

As a side effect, also fixes #8946 (sorry about that).