st-texture-cache: Add an async image-loading method that provides a callback that supplies the actor, which has been explicitly made the same size as the loaded image.
This is can be useful vs load_uri_async, which immediately returns
an empty actor that gets filled in later.  The problem with this
is, the actor is forced to obey the allocations set by the container,
which may not want to allow the full natural size of the image.  This
causes the image to be scaled down so it can be shown within the
allocated space.