cinnamon-settings: Fix hiding of the display panel's monitor labels when navigating back to the icon view.
When the display module was converted into a hybrid python/c module,
the mechanism for making sure the monitor labels get hidden was broken.

This adds a navigate-away hook to the module that will hide the c widget,
which in turn will cause the display module to hide the labels (it monitors
its own visibility already to manage the labels).