notification system: notification UI reworks
All code related to expanding and collapsing notifs has been
condensed down into the functionality we actually use. We have
been forcing the expand path for years, so it no longer exists
as a separate path.

The custom banner box has been removed. The ellipsization was
broken for unknown reasons, and it was a significant amount of
code to basically display short notifications on one line.

Support for custom actors in notifications is removed. The only
(internal)  user was unused code that I also removed in a previous
commit of this branch.