If statement appears to serve no purpose
This is my first attempt to make a change to someone else's code so feel free to give me feedback.
These 3 lines don't seem to have a purpose that I can see and I've checked at least 10 popular themes and they all work fine without them. Also, they appear to cause a bug: If the style value "height:" is set in the node of the css theme file, it makes no difference anyway, the height value is set to 30 regardless (I'm not sure way this is.) So If the text size (set in the css file) is too large, it disappears of the bottom - this is how I discovered the bug.