menu applet: search for words inbetween
The search function does a search only at the beginning of a word.
E.g. search for "Manager" does find "Update Manager", but searching for
"date" does not find "Update Manager" or a better example, searching for
"office" does not find "LibreOffice".

This commit (magic trick from mtwebster) removes the restriction to the beginning of a word.