Proper fix for builds with networkmanager disabled.
We've never linked to libnm, merely checked if it existed. During the
meson port, this somehow found its way into being over-linked into
cinnamon itself, resulting in #9727, which was improperly fixed in #9729
by only performing the over-linking if networkmanager is enabled.

Nevertheless, it remains overlinking. Remove it entirely, and comment
the dependency() call to document why we're looking for it but not
trying to link it to anything.