Make the menu applet resizeable.
This adds a class in js/ui/applet.js that can be attached to an applet
to allow it to be resized using the mouse.

This is currently only implemented for the menu applet but can be expanded
in the future.

The layout for the applet was cleaned up to allow proper functioning of
the three scroll view actors.

Minor issues:

The extra allocation for the scrollbars don't go away even though the
scrollbar itself is invisible. Because of this maybe we want to pad the
category labels with a few trailing spaces to counter the space lost from
the new scrollbar.

Also, in certain panel positions, resizing in one direction may result in
changes to the position of the opposite edge due to how the panel calculates
popup menus. This doesn't affect function, and in reality probably won't be
an issue for many people.