New ICON_NAMES Entries for Tilix terminal actions
Make icons in the context menu for Tilix (terminal emulator) reflect the behavior of those actions. The "New Session" action now uses the "tab-new-symbolic" icon, as a session in Tilix is equivalent to a tab, and the button within Tilix that creates a new session uses "tab-new-symbolic" as well. The "Preferences" action will now use the "preferences-other" icon, just like the "Prefs" entry following it in the list. Before, these two actions were default icons, and did not represent their respective actions. First time contributing, hope this is helpful/correct.