Attempt to fix clipping issue with osdWindow.js
https://github.com/linuxmint/cinnamon/issues/9758

I have literally zero idea what I am doing or how to code JavaScript, but my issue has been open more than 20 days with no response, so I am desperate at trying to get this resolved. I figure this is the only way to get someone to look at it?

This fixes the issue I described and also makes the Volume % font less blurry and more "crisp". There is some sort of issue with the OSD_SIZE and this._sizeMultiplier not playing nice with each other according to the previous values.