cinnamon-app.c: Have cinnamon_app_get_id() return the app's lookup key for the CinnamonAppSystem app table.
Some applets (menu, gwl and panel-launchers) want to use this id
for lookups at a later time. In the case of desktop files inside of
subfolders, cinnamon-menus prepends the parent path to the desktop
filename, and this is used as a key in CinnamonAppSys. This won't
match the id returned by g_app_info_get_id() causing app-sys lookups
to fail later on.

ref: https://github.com/linuxmint/mint20.1-beta/issues/40