Fix for desktop files in subfolders: use the desktop id generated in cinnamon-menus
Some applets (menu, gwl and panel-launchers) want to use this id
for lookups at a later time. In the case of desktop files inside of
subfolders, cinnamon-menus prepends the parent path to the desktop
filename, and this is used as a key in CinnamonAppSys. This won't
match the id returned by g_app_info_get_id() causing app-sys lookups
to fail later on.