various: Fix hidpi change detection.
This was broken with fractional scaling - we have traditionally
equated the changing of Xft.dpi to mean the scale changed, so we
could update our ui (cinnamon runs in 1x always, but we scale our
elements to match the global scale).

When fractional scaling is disabled (switched off thru the gui),
these old methods are fine - changes to the global scale affect
the dpi as well.

With it enabled, however, the following differences manifest:
- The global scale becomes only the base scale factor from which
  the fractional scale will be achieved.
- Only changes to the fractional value will affect the Xft.dpi
  value; changing the global scale does not.

Unfortunately cinnamon (as well as some apps in certain situations)
need to know what the global scale is, and when it changes, so the
dpi is no longer a reliable value to monitor.

This uses the primary GdkMonitor to listen for changes to its
scale-factor property and notify the appropriate parties when it
changes.