Fix #9822, another GWL issue
When the "Show windows from all workspaces" setting is enabled, the Grouped Window List does not clear app instances from other workspaces when they are closed. This issue seems to occur only when the app is moved to a  workspace that has not been viewed yet (eg. using the title bar menu option "Move to Another Workspace")

This isn't the most elegant way to fix this issue, but I've tried to do it with little impact on existing code.