power applet: Fix _devicesChanged so that calls to csd-power are handled sequentially.
GetPrimaryDeviceRemote would be called, followed immediately by
a call to GetDevicesRemote - without waiting for the first call
to complete. Since the second call's callback is expecting info
received in the first's callback, if makes more sense (and is
more clear) to nest the second call in the repsonse of the first.

This really shouldn't change actual behavior since csd-power responds
to these calls synchronously and call order is preserved by dbus
(so we'll have a response from the first call before we get a
response to the second).

See:
https://bugs.freedesktop.org/show_bug.cgi?id=101563