appSwitcher3D.js: Don't fail completely if the window actor is gone when removing previews.
The clone of a window can be kept alive independently of the actor
it's cloning (especially due to language bindings) so we should
check if we have it prior to trying to access it.

Added a try/catch block for hiding the previews in case anything
else fails, cinnamon won't be left in an unusable state.

ref: #9172