Fix cinnamon-xlet-makepot so it's always gives the same output
The issue exists os.walk's list elements are in arbitrary order [link](https://stackoverflow.com/a/18282401). Because of that the resulting .pot file's strings will be in different order on different systems, making reviewing PRs for them hard to track.

Sorting os.walk's output in every place fixes this issue.