window list: Keep track of window positions in the list so the order can be restored between restarts.
A generic setting is used to store a sequence of window XIDS that
correspond to the order of each window in the list. This gets updated
any time a window is added or removed.

This isn't intended to be any sort of window 'preset' feature - it's
only to track current windows - every window will always have a
different xid.

This could be ported to gwl once any issues are worked out here
(it shouldn't be too hard to do, other than extra consideration for
pinned apps and grouped/ungrouped modes).