cinnamon-xlet-makepot: Add header and metadata to .pot file
If we only have `.js` files with translations, xgettext is never
triggered and we end up with a .pot file without a header and without
metadata.

Also if xgettext is not triggered, the old .pot file is never deleted
and old unused strings are never deleted. Translators have to translate
strings for nothing.
Therefore we need to delete existing .pot files at the very beginning.

Fixes https://github.com/linuxmint/cinnamon/issues/9961