cs_startup.py: Don't write an invalid override desktop file the second time it's created in the same process.
The first time a new desktop file is created, we base whether or
not we need to create one on self.user_position being set or not.
If we're reverting back to the system file, we need to clear that
property so we correctly refer to the system file again if app is
toggled a second time.

This only affects apps that have their own autostart desktop files
to begin with. User-created ones aren't affected because they're
being modified in-place.

Fixes linuxmint/xapp#129.