Remove window-close action on Alt-Tab Q shortcut
Fixes #9787

Closing windows like that (especially with the Q key, which is dangerously close to the Tab key that the user often intends to press) is a destructive and not undoable action, which makes for a bad user experience when it happens. As per https://github.com/linuxmint/cinnamon/pull/10029#issuecomment-813109829 it seems like the feature might have low enough usage that removing it is better than maintaining it and having a separate configuration option.