Add icon for "all applications"
The lack of icon for this category button caused the button to be less tall than the other ones, and messed up the category text alignment. It also felt awkward compared to the rest.

This commit adds a category icon so the menu no longer triggers my OCD. See: *happy perfectionist noises*