xapp status applet: update the tooltip text immediately
Previously, the tooltip wasn't updated until the next time it was shown, meaning that if the value changed while the tooltip was already visible, it would continue to display the outdated text until the mouse was moved away and then back again. This could be a problem in situations where the user scrolls on the icon to change a value (eg. volume) which is displayed in the tooltip.