lookingGlass.js: delay and aggregate window tracker updates
For some reason, windows still exist in the list by the time our
update happens when reacting to a MetaWindow::unmanage signal.
Delaying the update for 200ms seems to completely fix this issue
in my configuration. In addition, the delay is set to reset any time
a new update is queued during the delay so that multiple consecutive
update signals can be aggregated into one. This may reduce the CPU
time when many windows open or close at once.