org.cinnamon.gschema.xml: Add libreoffice to focus passthru list.
LibreOffice fails to set an wm_class immediately when creating
a new window, so it defautls to the process name (soffice), but
if an already-opened document is 'opened' again, the existing
window's wm_class will be 'libreoffice-writer' or whatever specific
app the window is for, and also needs help bringing focus to that
activated window.

Note: by design LibreOffice doesn't focus new windows (other than
the first), but we want to keep the existing behavior from 4.8. See:
https://github.com/linuxmint/cinnamon/issues/10202#issuecomment-879220661