main.js: Fix activateWindow so it will respect the workspace index if provided.
The index argument isn't used much, but expo does use it, for
activating a window on a different workspace.

This isn't quite back to original behavior but in places I've seen
it in spices, and in the rest of cinnamon, there will be little
difference, and this makes the api more useful.

Fixes #10253