Fix popup menu positioning along the bottom edge of the workarea (when the menu's orientation hasn't been set).
Orientation is set by the appletManager, any other use does not
it, so 'flipping' may not occur even when necessary.

Changed the final position in this scenario so that the menu is
up against the edge of the workarea, rather than flipped and
positioned against the pointer. This brings it in line with
the behavior for the right edge of the workarea.