Use muffin's sound player instead of csd-sound.
This eliminates the need for csd-sound.

Changed the volume slider and sound applet's mouse-scrolling to
filter out 'smooth' scroll events, as we were playing sounds for
two events at a time (we received both smooth and directional
scroll events for the same action).

Removed any volume control for sounds - it was only used in the
sound applet, and was never actually used, as stream.decibal would
usually be a negative value. It's unnecessary anyhow, as the
volume sound will play at the current volume.. that is being
adjusted.