calendar applet: Add support for evolution-data-server (#10453)
* calendar applet: Add goa calendar support.

* control: add dependencies for cinnamon-calendar-server.

* Implement periodic updates instead of push-style.

It may not be desirable to have a 60mb process running in
the background to track events that mostly never change.

If enabled, the server will be started on-demand (if the
user changes months) and on a timer in the applet, to start
and refresh events every 15 minutes by default. It then
exits after 20s.