appletManager.js: Fix highlighting when adding an applet to a panel.
There are two paths an applet takes on its way to its
panel, depending on whether or not the applet has been
previously loaded from disk.

We reload all applets in a given panel zone whenever
their layouts change. When re-adding them, be more
selective about what's a new instance, and one that
was already there, but some part of its definition
may have changed (like its position on the panel).

Trigger the flash also in the asynchronous loading
callback, which we can assume is the user adding a
previously not-loaded applet, as long as we're not
performing the initial load of all applets during
startup.