Simplify reverse scrolling implementation
Reverse mouse wheel scrolling direction should not affect
left click cycle app action. This simplification sets
the left click direction to be "right" instead of "down",
in that way there is no need to check if the action is left click
because reverse scrolling only affects "up" and "down" directions.