Remove additional trigger for updateSeparatorVisibility.
This was triggering a crash with gwl when opening an active
window's popup menu. This is because the popup was being
constructed *after* the menu open-state-changed signal,
which is wrong.

The gwl issue is fixed by having the menu constructed before
showing, but there may be potential for this in 3rd-party
xlets.

ref:
https://github.com/linuxmint/cinnamon/commit/1d07cf016ab68c7f21bba5f392ff158bd2907502