Fix a couple typos in latest code (#10569)
* Fix a couple typos in latest code

In cs_sound.py, a missing underscore causes an undefined function to
be called. The fix is to change it to set_x_align.

In accessibility.js, there is a missing break in the switch statement.

In cr-utils.c, out_index + 4 is used twice, the second should have been
out_index + 5.