Fix the XApp Status Applet tooltip placement on vertical panels
When the XApp Status Applet is placed on a vertical panel the location
of the tooltip is always in the same place (middle of the applet box
rather then aligned with the app status icon that the mouse is hovering
over). This fix uses a unique Tooltip instance for each status app icon
so that the tooltip will be properly aligned with the icon in all cases.