panel.js: Fix autohide and intellihide (partially).
On the edge of the stage we always get enter->leave->(maybe)enter
events, causing the panel to be jumpy when we rely on these events
to show or hide the panel.

This assists by comparing the event's position against the
position of the 1px activation strip, and discards any leave events
that occur over that strip.

Fixes #10869.