keybindings: Fix arguments sent by invoke_keybinding_action_by_id.
This is called in main.js when in a modal state (_stageEventHandler).
The callback expects 4 arguments - 3 from muffin and the callback
data (in this case the media action id). When we invoke it here
artificially we have to use the same number of arguments.

Fixed some logging useful for debugging.