windowAttentionHandler: remove filtering (no longer necessary)
Retested all of the 'passthru' apps we had, none of them appear
to misbehave any longer.

If there's no window list, this will send a notification only
now, which the user can click to activate the window. Otherwise
it does nothing.

It adds a new role for an attention handler, but also checks loaded
applets for keywords so this doesn't break any existing applets.

This fixes firefox focusing AND flashing for attention when clicking
a link from another program.