panels: Fix relayout problem when a monitor is removed or added.
Calling global.display.xinerama_index_to_logical_index() will
return 0 if the x index doesn't exist, which causes that panel
to think it should be put on monitor 0 (instead of the actual
monitor 0 panel), depending on their order in the settings array.

Check if the panel def's monitor is higher than the number of
monitors, and leave it unchanged (this will make the panel manager
skip it because the monitor index is out of range).