extensions: Allow an applet to assume more than one 'role'.
f4b5233998f simplified the windowAttentionHandler to have it only
function as an 'attention grabber'  if no other extensions were
handling this. To do this, a new xlet role was added -
'window-attention-handler', which would apply to any window list
applets (because they all provide this).

This caused a regression in GWL, as it already claimed the
'panel-launcher' role (which I overlooked), and adding the new
role overwrote that one, breaking the menu applet's 'add to panel'
feature.