g-w-l/workspace-switcher: Fix random crash and errors when adding and removing workspaces.
gwl: Remove workspaces last-to-first when removing so the removal
of lower-index workspaces won't affect the position of higher
ones while they're being destroyed.

workspace-switcher: Connect to 'workspaces-reordered', not
'workareas-changed', and perform a button refresh immediately
when receiving these signals.

The workareas-changed signal only fires from one place that the
n_workspaces property notify is not, and it's also on an idle
callback inside of muffin, so by the time workareas-changed got
sent, the situation could have already changed, plus it would be
the second signal received for the change (n_workspaces being
updated immediately).

The workspaces-reordered is called in that single instance that
n_workspaces is not, so connect to that instead, since it's also
immediate.

ref:
linuxmint/muffin@603dba89297695a53fe9
https://github.com/linuxmint/mint21-beta/issues/89