cs_default.py: Add entry for terminal exec-arg
Terminals have options to provide it with a command to run, which may
vary between different terminals. The option text is saved in the
`org.cinnamon.desktop.default-applications.terminal` GSettings schema,
and used in various places in the Cinnamon desktop. This was previously
not configurable in a GUI and will require the user to enter a gsettings
command, so add the functionality to "Preferred Applications".

Fixes #10938.