Reimplement the locate pointer functionality
We lost this in the rebase due to changes in how they are handled.
Reimplement them the new way and also allow them to be themed by the
Cinnamon theme. Also adding a new ripples.js that we can reuse for the
hot corners since this same functionality is now being used in more than
one place.