grouped-window-list: Don't create a new clone each time getThumbnail is called, don't connect to notify::size.
getThumbnail was getting called a lot, as notify::size would be
signaled many times during an effect. Working off the underlying
texture is better - it only resizes when the actual client content
does, not while the actor is being animated.

We still want to continue refreshing the thumbnail while it remains
static only (and not a true clone), otherwise it will never end
up with a live preview.