expo: Don't create monitor clones until the ExpoThumbnailBox is allocated.
Creating them before caused the box to be immediately allocated
(to accomodate the clones) which caused its allocation signal
to be sent before we connected to it to begin the primary
animation. This was also causing random actor map/unmap issues,
leading to potential crashes.

Disable all expo effects when desktop effects are disabled.